requirements.txt
PyQt5==5.15.9
numpy==1.21.6
pykdtree==1.3.3
opencv-python-headless==4.8.0.76
tensorboard==2.14.0
dominate==2.8.0
PyMCubes==0.1.4
Cython==3.0.0
pyOpenGL==3.1.7
trimesh==3.23.3
easydict==1.10
open3d==0.17.0
openmesh==1.1.0

